# Use a cache on the backend itself to store the context.