Fixes #208