@hameerabbasi 2019-09-27
This is still WIP, needs `tp_visit` on both context managers.

### Benchmarks
#### Case 1
```python
with ua.set_backend(Backend):
    pass
```

#### Case 2
```python
with ctx:
    pass
```

#### `master`
Case 1: 409 ns ± 7.46 ns
Case 2: 92.5 ns ± 1.02 ns

#### `ctx-cache`
Case 1: 258 ns ± 3.27 ns
Case 2: 88.4 ns ± 1.74 ns